alignment_agent/requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Natural Language Processing
spacy>=3.4.0
nltk>=3.7
transformers>=4.20.0
sentence-transformers>=2.2.0
torch>=1.12.0

# LLM Integration
openai>=1.0.0
anthropic>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.20

# Knowledge Graph and Vector Database
networkx>=2.8.0
neo4j>=5.0.0
chromadb>=0.4.0
faiss-cpu>=1.7.0
rdflib>=6.2.0
py2neo>=2021.2.3

# IFC Processing
ifcopenshell>=0.7.0
# pythonOCC-core>=7.6.0  # Optional: 3D geometry processing
# pythreejs>=2.4.0      # Optional: 3D visualization

# Text Processing and Document Handling
PyPDF2>=3.0.0
python-docx>=0.8.11
markdown>=3.4.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Data Serialization and Configuration
PyYAML>=6.0
json5>=0.9.0
toml>=0.10.2
python-dotenv>=1.0.0

# Logging and Monitoring
loguru>=0.6.0
rich>=12.0.0
tqdm>=4.64.0

# Web Framework (for API endpoints)
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=2.0.0

# HTTP and Networking
requests>=2.28.0
httpx[socks]>=0.24.0
aiohttp>=3.8.0

# Utilities
click>=8.1.0
typer>=0.9.0
dateutils>=0.6.12
fuzzy>=1.2.2
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

# Data Processing
pyarrow>=10.0.0
openpyxl>=3.0.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Development Tools
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
isort>=5.10.0
pre-commit>=2.20.0

# Performance and Optimization
numba>=0.56.0
joblib>=1.2.0

# Async Support
aiofiles>=22.1.0

# Environment and Configuration
python-multipart>=0.0.6
watchdog>=2.2.0